# produits/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Product API docs",
        "description": "This is Product API documentation",
        "contact": {
            "email": "yassine.mechiche@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/microservices/produits/public"
        }
    ],
    "paths": {
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get list of products",
                "description": "Get list of products",
                "operationId": "listProducts",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/listProductResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity (validation failed)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorValidation"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "General server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Add a product",
                "description": "Add a product",
                "operationId": "AddProduct",
                "requestBody": {
                    "description": "Add a product",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nom",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Description",
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    },
                                    "quantity_in_stock": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Basic"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity (validation failed)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorValidation"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "General server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product details",
                "description": "Get product details",
                "operationId": "productDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/detailProductResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity (validation failed)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorValidation"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "General server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update a product",
                "description": "Update a product",
                "operationId": "UpdateProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Update a product",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nom",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Description",
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    },
                                    "quantity_in_stock": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Basic"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity (validation failed)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorValidation"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "General server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete a product",
                "description": "Delete a product",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Basic"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity (validation failed)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorValidation"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "General server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ApiResponseError": {
                "required": [
                    "base_url",
                    "status",
                    "message",
                    "status_code"
                ],
                "properties": {
                    "base_url": {
                        "type": "string",
                        "format": "url"
                    },
                    "status": {
                        "type": "boolean",
                        "default": "false"
                    },
                    "status_code": {
                        "type": "integer"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ApiResponseErrorValidation": {
                "required": [
                    "base_url",
                    "status",
                    "message",
                    "status_code"
                ],
                "properties": {
                    "base_url": {
                        "type": "string",
                        "format": "url"
                    },
                    "status": {
                        "type": "boolean",
                        "default": "false"
                    },
                    "status_code": {
                        "type": "integer"
                    },
                    "message": {
                        "type": "string"
                    },
                    "errors": {
                        "properties": {
                            "field_one": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            },
                            "field_two": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ApiResponseSuccess": {
                "required": [
                    "base_url",
                    "status",
                    "message"
                ],
                "properties": {
                    "base_url": {
                        "type": "string",
                        "format": "url"
                    },
                    "status": {
                        "type": "boolean",
                        "default": "true"
                    },
                    "status_code": {
                        "type": "integer"
                    },
                    "message": {
                        "type": "string",
                        "example": "OK"
                    },
                    "data": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "Basic": {
                "required": [
                    "base_url",
                    "status",
                    "message"
                ],
                "properties": {
                    "base_url": {
                        "type": "string",
                        "format": "url"
                    },
                    "status": {
                        "type": "boolean"
                    },
                    "status_code": {
                        "type": "integer"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "detailProductResponse": {
                "required": [
                    "base_url",
                    "status",
                    "message",
                    "data"
                ],
                "properties": {
                    "base_url": {
                        "type": "string",
                        "format": "url"
                    },
                    "status": {
                        "type": "boolean"
                    },
                    "status_code": {
                        "type": "integer"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/ProductModel"
                    }
                },
                "type": "object"
            },
            "listProductResponse": {
                "required": [
                    "base_url",
                    "status",
                    "message",
                    "data"
                ],
                "properties": {
                    "base_url": {
                        "type": "string",
                        "format": "url"
                    },
                    "status": {
                        "type": "boolean"
                    },
                    "status_code": {
                        "type": "integer"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductModel"
                        }
                    }
                },
                "type": "object"
            },
            "ProductModel": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "quantity_in_stock": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Products",
            "description": "Products routes"
        }
    ],
    "security": [
        []
    ]
}